+++ Example/ExampleModel.cs
@@ -3,7 +3,7 @@ using BigExcelCreator.Enums;
 
 namespace Example
 {
-    [ExcelHeaderStyleFormat(6)]
+    [ExcelHeaderStyleFormat("bold center")]
     internal class ExampleModel
     {
         [ExcelColumnName("Identifier")]
@@ -40,13 +40,14 @@ namespace Example
         [ExcelColumnOrder(4)]
         [ExcelColumnType(CellDataType.Number)]
         [ExcelColumnWidth(20)]
+        [ExcelConditionalFormatFormula("RED", "E2<6100")]
         public decimal Salary { get; set; }
 
         [ExcelColumnName("Monthly Bonus")]
         [ExcelColumnOrder(5)]
         [ExcelColumnType(CellDataType.Number)]
         [ExcelColumnWidth(20)]
-        [ExcelStyleFormat(8)]
+        [ExcelStyleFormat("bold italic center")]
         public double? Bonus { get; set; }
 
         [ExcelColumnName("Monthly Net Income")]